Add Snakemake config parsing file
This file is expected to hold functions for parsing Snakemake workflow configs. 
It currently only includes the `resolve_config_path` function that is needed to 
support external analysis directories with `nextstrain run`. 

The `resolve_config_path` function is built on the work @tsibley did in 
measles and the work @jameshadfield did in avian-flu.

<https://github.com/nextstrain/measles/blob/efda4428d83bf0de092db21d2dd00822a0a41509/shared/functions.smk#L3>
<https://github.com/nextstrain/avian-flu/blob/71f005f94f4256e959c90a2830c646a660137a75/rules/config.smk#L26>