scripts/assign-colors: only use 'focal' column if dtype is `bool`
If the Series is not a boolean, then pd.loc will treat it as an index and 
raise a KeyError, e.g.

```
KeyError: "None of [Index(['True', 'foo'], dtype='object')] are in the [index]"
```

I thought about forcing the dtype to bool with `.astype('bool')`, but that felt 
more surprising than just checking the column dtype.