Add resolve_filepaths() and helper functions
This code is already used in measles¹ and is expected to be used by other
repos.

I placed the functions above resolve_config_path() because, while
resolve_config_path() is used directly in Snakemake code, eventually it
may be replaced by calls to resolve_filepaths() which uses it as a
helper function.

¹ https://github.com/nextstrain/measles/blob/db993ffa3a0ab80bd46d68f5e30ee47023525bb8/phylogenetic/rules/config.smk#L29-L94